# feat: implement parallel streaming output rails execution