@Pouyanpi 2025-07-10
# PERFORMANCE COMPARISON SUMMARY

## Config 1

_config:_
  - Stream first: False
  - Chunk size: 200
  - Context size: 50
  
_Output Rails:_
  **- content safety check output $model=content_safety**
  - self check output



**Latency Improvements (First Chunk):**
  Test 1: 2.488s → 1.535s (38.3% faster, 1.62x speedup)
  Test 2: 2.255s → 0.956s (57.6% faster, 2.36x speedup)
  Test 3: 2.097s → 0.567s (73.0% faster, 3.70x speedup)

**Total Duration Comparison:**
  Test 1: 2.489s (sequential) vs 2.117s (parallel)
  Test 2: 2.256s (sequential) vs 1.525s (parallel)
  Test 3: 2.099s (sequential) vs 1.097s (parallel)

**Rail Execution Details:**
  - Total LLM calls: 30
  - Rails executed:
    - content_safety_check_output $model=content_safety: 12 call(s)
    - general: 6 call(s)
    - self_check_output: 12 call(s)


## Config 2 (Lower chunk size higher output rails invocation)
_config:_

  - Streaming enabled: True
  - Stream first: False
  - Chunk size: 30
  - Context size: 5

_Output Rails:_
  - content safety check output $model=content_safety
  - self check output
  
  
**Latency Improvements (First Chunk):**
  Test 1: 2.015s → 0.690s (65.7% faster, 2.92x speedup)
  Test 2: 1.605s → 0.586s (63.5% faster, 2.74x speedup)
  Test 3: 1.547s → 0.366s (76.3% faster, 4.23x speedup)

**Total Duration Comparison:**
  Test 1: 3.925s (sequential) vs 2.202s (parallel)
  Test 2: 4.507s (sequential) vs 1.162s (parallel)
  Test 3: 1.547s (sequential) vs 0.940s (parallel)


**Rail Execution Details:**
  - Total LLM calls: 44
  - Rails executed:
    - content_safety_check_output $model=content_safety: 19 call(s)
    - general: 6 call(s)
    - self_check_output: 19 call(s)